requirements.txt
openai
fhirpathpy
flask
python-dotenv
gunicorn
flask
requests
fhir.resources
ollama